refactor: split backend/ffmpeg_tools.py (1301 lines) into package
Decomposed monolithic ffmpeg_tools.py into 6 focused modules:
- discovery.py: binary finding, version validation, repair/install
- probe.py: video metadata probing via ffprobe
- color.py: colour-space filter chain (build_exr_vf + helpers)
- extraction.py: hwaccel detection, EXR extraction, DWAB recompression
- stitching.py: image sequence to video
- metadata.py: sidecar JSON read/write

Updated test_ffmpeg_tools.py to use package imports.
All 427 tests pass.
